fixed bugs in registertation

+++ app/Http/Controllers/Auth/RegisterController.php
@@ -79,9 +79,10 @@ class RegisterController extends Controller
                               'regex:/[0-9]/', 
                               'confirmed',
        ],   
-            /* Added by Achyut Neupane */
-            'referralemail' => 'required|exists:emails,email'
+            // /* Added by Achyut Neupane */
+        //    'referralemail' => 'required|exists:emails,email'
         ]);
+
         $user = User::create([
             'name' => $request->name,
             'gender' => $request->gender,
@@ -91,11 +92,18 @@ class RegisterController extends Controller
             'password' => Hash::make($request->password),
             'verification_code' => sha1(time())
         ]);
+        
+        /* if someone refer */
+        if($request->referralemail){
+ 
         $referral = Email::with('user')->where('email',$request->referralemail)->first()->id;
         $refer = Refer::where('referred_by',$referral)->where('email',$request->email)->first();
         $refer->registered = true;
         $refer->save();
         Refer::where('email',$request->email)->where('registered',false)->delete();
+
+        }
+        
 
         $user->emails()->create([
             'email' => $request->email,
